+++ packages/deep-engine/src/runtimePackage/index.ts
@@ -10,6 +10,7 @@ export * from "./coordinates.js";
 export { buildDashboardRuntimePackage } from "./dashboard.js";
 export { buildChartRuntimePackage } from "./chart.js";
 export { buildDashboardCompositionRuntimePackage } from "./dashboardComposition.js";
+export { dashboardRuntimePageId } from "./dashboardAuthorIdentity.js";
 export * from "./dashboardCompositionTypes.js";
 export { DashboardCandidateController } from "./dashboardCandidateController.js";
 export type * from "./dashboardCandidateTypes.js";